requirements-4.txt
python-telegram-bot==20.3
aiohttp==3.9.1
python-dotenv==1.0.0
qrcode[pil]==7.4.2
Pillow==10.1.0
telethon==1.36.0
